requirements.txt
Django==4.2.7
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.24.3
Pillow==10.1.0
